+++ docs_src/connections/contextmanager.py
@@ -1,4 +1,4 @@
-from edgy import Registry, Instance, monkay
+from edgy import Registry, Instance, monkay, run_sync
 
 models = Registry(database="sqlite:///db.sqlite", echo=True)
 
@@ -6,9 +6,9 @@ models = Registry(database="sqlite:///db.sqlite", echo=True)
 # load settings
 monkay.evaluate_settings(ignore_import_errors=False)
 # monkey-patch app so you can use edgy shell
-monkay.set_instance(Instance(registry=registry))
+monkay.set_instance(Instance(registry=models))
 
 
-def main():
+def main() -> None:
     with models.with_async_env():
-        edgy.run_sync(User.query.all())
+        run_sync(User.query.all())