+++ docs_src/connections/contextmanager_with_loop.py
@@ -24,7 +24,7 @@ def get_application():
     # load settings
     monkay.evaluate_settings(ignore_import_errors=False)
     # monkey-patch app so you can use edgy shell
-    monkay.set_instance(Instance(registry=registry, app=app))
+    monkay.set_instance(Instance(registry=models, app=app))
     return app