+++ docs_src/connections/django.py
@@ -1,7 +1,7 @@
 from django.core.asgi import get_asgi_application
 
 
-from edgy import Registry, Instance
+from edgy import Registry, Instance, monkay
 
 models = Registry(database="sqlite:///db.sqlite", echo=True)
 
@@ -11,4 +11,4 @@ application = models.asgi(handle_lifespan=True)(get_asgi_application())
 # load settings
 monkay.evaluate_settings(ignore_import_errors=False)
 # monkey-patch app so you can use edgy shell
-monkay.set_instance(Instance(registry=registry, app=app))
+monkay.set_instance(Instance(registry=models, app=application))