+++ docs_src/connections/manual.py
@@ -7,7 +7,7 @@ async def main():
     # load settings
     monkay.evaluate_settings(ignore_import_errors=False)
     # monkey-patch app so you can use edgy shell
-    monkay.set_instance(Instance(app=app, registry=registry))
+    monkay.set_instance(Instance(registry=models))
     await models.__aenter__()
     try:
         ...