+++ edgy/core/connection/asgi.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from collections.abc import Awaitable, Callable
+from collections.abc import Awaitable, Callable, MutableMapping
 from contextlib import suppress
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any
@@ -10,9 +10,9 @@ if TYPE_CHECKING:
 
 ASGIApp = Callable[
     [
-        dict[str, Any],
-        Callable[[], Awaitable[dict[str, Any]]],
-        Callable[[dict[str, Any]], Awaitable[None]],
+        MutableMapping[str, Any],
+        Callable[[], Awaitable[MutableMapping[str, Any]]],
+        Callable[[MutableMapping[str, Any]], Awaitable[None]],
     ],
     Awaitable[None],
 ]
@@ -30,14 +30,14 @@ class ASGIHelper:
 
     async def __call__(
         self,
-        scope: dict[str, Any],
-        receive: Callable[[], Awaitable[dict[str, Any]]],
-        send: Callable[[dict[str, Any]], Awaitable[None]],
+        scope: MutableMapping[str, Any],
+        receive: Callable[[], Awaitable[MutableMapping[str, Any]]],
+        send: Callable[[MutableMapping[str, Any]], Awaitable[None]],
     ) -> None:
         if scope["type"] == "lifespan":
             original_receive = receive
 
-            async def receive() -> dict[str, Any]:
+            async def receive() -> MutableMapping[str, Any]:
                 message = await original_receive()
                 if message["type"] == "lifespan.startup":
                     try: